feat: 사용자 닉네임 표시
-  마이페이지에 있던 getUserInfo 로직을 useAuth로 이동
